Enable peer-to-peer exchange between processes on the same GPU when MPS is enabled (this should give a good speedup to using MPS).  Moved comm_peer2peer_init into comm_common.c frather than rely on MPI and QMP variants.
