+++ lib/comm_mpi.cpp
@@ -42,12 +42,21 @@ struct MsgHandle_s {
 static int rank = -1;
 static int size = -1;
 static int gpuid = -1;
-static bool peer2peer_enabled[2][4] = { {false,false,false,false},
-                                        {false,false,false,false} };
-static bool peer2peer_init = false;
 
 static char partition_string[16];
 static char topology_string[16];
+
+
+void comm_gather_hostname(char *hostname_recv_buf) {
+  // determine which GPU this rank will use
+  char *hostname = comm_hostname();
+  MPI_CHECK( MPI_Allgather(hostname, 128, MPI_CHAR, hostname_recv_buf, 128, MPI_CHAR, MPI_COMM_WORLD) );
+}
+
+void comm_gather_gpuid(int *gpuid_recv_buf) {
+  MPI_CHECK(MPI_Allgather(&gpuid, 1, MPI_INT, gpuid_recv_buf, 1, MPI_INT, MPI_COMM_WORLD));
+}
+
 
 void comm_init(int ndim, const int *dims, QudaCommsMap rank_from_coords, void *map_data)
 {
@@ -74,14 +83,13 @@ void comm_init(int ndim, const int *dims, QudaCommsMap rank_from_coords, void *m
   comm_set_default_topology(topo);
 
   // determine which GPU this MPI rank will use
-  char *hostname = comm_hostname();
   char *hostname_recv_buf = (char *)safe_malloc(128*size);
 
-  MPI_CHECK( MPI_Allgather(hostname, 128, MPI_CHAR, hostname_recv_buf, 128, MPI_CHAR, MPI_COMM_WORLD) );
+  comm_gather_hostname(hostname_recv_buf);
 
   gpuid = 0;
   for (int i = 0; i < rank; i++) {
-    if (!strncmp(hostname, &hostname_recv_buf[128*i], 128)) {
+    if (!strncmp(comm_hostname(), &hostname_recv_buf[128*i], 128)) {
       gpuid++;
     }
   }
@@ -108,103 +116,6 @@ void comm_init(int ndim, const int *dims, QudaCommsMap rank_from_coords, void *m
   snprintf(partition_string, 16, ",comm=%d%d%d%d", comm_dim_partitioned(0), comm_dim_partitioned(1), comm_dim_partitioned(2), comm_dim_partitioned(3));
   snprintf(topology_string, 16, ",topo=%d%d%d%d", comm_dim(0), comm_dim(1), comm_dim(2), comm_dim(3));
 }
-
-void comm_peer2peer_init(const char* hostname_recv_buf)
-{
-  if (peer2peer_init) return;
-
-  bool disable_peer_to_peer = false;
-  char *enable_peer_to_peer_env = getenv("QUDA_ENABLE_P2P");
-  if (enable_peer_to_peer_env && strcmp(enable_peer_to_peer_env, "0") == 0) {
-    if (getVerbosity() > QUDA_SILENT) printfQuda("Disabling peer-to-peer access\n");
-    disable_peer_to_peer = true;
-  }
-
-  // disable peer-to-peer comms in one direction if QUDA_ENABLE_P2P=-1
-  // and comm_dim(dim) == 2 (used for perf benchmarking)
-  bool disable_peer_to_peer_bidir = false;
-  if (enable_peer_to_peer_env && strcmp(enable_peer_to_peer_env, "-1") == 0) {
-    if (getVerbosity() > QUDA_SILENT) printfQuda("Disabling bi-directional peer-to-peer access\n");
-    disable_peer_to_peer_bidir = true;
-  }
-
-  if (!peer2peer_init && !disable_peer_to_peer) {
-
-    // first check that the local GPU supports UVA
-    cudaDeviceProp prop;
-    cudaGetDeviceProperties(&prop,gpuid);
-    if(!prop.unifiedAddressing) return;
-
-    comm_set_neighbor_ranks();
-
-    char *hostname = comm_hostname();
-
-    int *gpuid_recv_buf = (int *)safe_malloc(sizeof(int)*size);
-
-    // There are more efficient ways to do the following,
-    // but it doesn't really matter since this function should be
-    // called just once.
-    MPI_CHECK( MPI_Allgather(&gpuid, 1, MPI_INT, gpuid_recv_buf, 1, MPI_INT, MPI_COMM_WORLD) );
-
-    for(int dir=0; dir<2; ++dir){ // backward/forward directions
-      for(int dim=0; dim<4; ++dim){
-	int neighbor_rank = comm_neighbor_rank(dir,dim);
-	if(neighbor_rank == rank) continue;
-
-	// disable peer-to-peer comms in one direction
-	if ( ((rank > neighbor_rank && dir == 0) || (rank < neighbor_rank && dir == 1)) &&
-	     disable_peer_to_peer_bidir && comm_dim(dim) == 2 ) continue;
-
-	// if the neighbors are on the same
-	if (!strncmp(hostname, &hostname_recv_buf[128*neighbor_rank], 128)) {
-	  int neighbor_gpuid = gpuid_recv_buf[neighbor_rank];
-	  int canAccessPeer[2];
-	  cudaDeviceCanAccessPeer(&canAccessPeer[0], gpuid, neighbor_gpuid);
-	  cudaDeviceCanAccessPeer(&canAccessPeer[1], neighbor_gpuid, gpuid);
-
-      // this was introduced with CUDA 8
-#if CUDA_VERSION >= 8000
-      int accessRank[2];
-      cudaDeviceGetP2PAttribute(&accessRank[0], cudaDevP2PAttrPerformanceRank, gpuid, neighbor_gpuid);
-      cudaDeviceGetP2PAttribute(&accessRank[1], cudaDevP2PAttrPerformanceRank, neighbor_gpuid, gpuid);
-#endif
-
-	  if(canAccessPeer[0]*canAccessPeer[1]){
-	    peer2peer_enabled[dir][dim] = true;
-	    if (getVerbosity() > QUDA_SILENT)
-	      printf("Peer-to-peer enabled for rank %d (gpu=%d) with neighbor %d (gpu=%d) dir=%d, dim=%d, performance rank = (%d, %d)\n",
-#if CUDA_VERSION >= 8000
-                  comm_rank(), gpuid, neighbor_rank, neighbor_gpuid, dir, dim, accessRank[0], accessRank[1]);
-#else
-                  // default to 0 for CUDA < 8
-                  comm_rank(), gpuid, neighbor_rank, neighbor_gpuid, dir, dim, 0, 0);
-#endif
-	  }
-	} // on the same node
-      } // different dimensions - x, y, z, t
-    } // different directions - forward/backward
-
-    host_free(gpuid_recv_buf);
-  }
-
-  peer2peer_init = true;
-
-  // set gdr enablement
-  if (comm_gdr_enabled()) {
-    printfQuda("Enabling GPU-Direct RDMA access\n");
-  } else {
-    printfQuda("Disabling GPU-Direct RDMA access\n");
-  }
-
-  checkCudaErrorNoSync();
-  return;
-}
-
-
-bool comm_peer2peer_enabled(int dir, int dim){
-  return peer2peer_enabled[dir][dim];
-}
-
 
 int comm_rank(void)
 {
