+++ lib/comm_single.cpp
@@ -3,6 +3,7 @@
  */
 
 #include <stdlib.h>
+#include <string.h>
 #include <csignal>
 #include <comm_quda.h>
 
@@ -15,15 +16,19 @@ void comm_init(int ndim, const int *dims, QudaCommsMap rank_from_coords, void *m
   comm_set_default_topology(topo);
 }
 
-void comm_peer2peer_init(const char *hostname_buf) {}
-
-bool comm_peer2peer_enabled(int die, int dim) { return false; }
-
 int comm_rank(void) { return 0; }
 
 int comm_size(void) { return 1; }
 
 int comm_gpuid(void) { return 0; }
+
+void comm_gather_hostname(char *hostname_recv_buf) {
+  strncpy(hostname_recv_buf, comm_hostname(), 128);
+}
+
+void comm_gather_gpuid(int *gpuid_recv_buf) {
+  gpuid_recv_buf[0] = comm_gpuid();
+}
 
 MsgHandle *comm_declare_send_displaced(void *buffer, const int displacement[], size_t nbytes)
 { return NULL; }
